Split up identity module into sub-modules.
